Spliting tests in one one test file: IdSearchTests.
